# perf(fake-tcp) faster checksum calculation