--- a/fake-tcp/Cargo.toml
+++ b/fake-tcp/Cargo.toml
@@ -10,6 +10,10 @@ description = """
 A TUN interface based, user space, asynchronous and high performance TCP stack that allows
 packet oriented tunneling with minimum overhead.
 """
+
+[features]
+benchmark = []
+
 [dependencies]
 bytes = "1"
 pnet = "0.28.0"
@@ -17,3 +21,4 @@ tokio-tun = "0.3.15"
 tokio = { version = "1.11.0", features = ["full"] }
 rand = { version = "0.8.4", features = ["small_rng"] }
 log = "0.4"
+internet-checksum = "0.2.0"